chore: consolidate stores under shared directory
