+++ web/apps/mfe-spectrogram/src/components/__tests__/FooterInteractions.test.tsx
@@ -44,7 +44,7 @@ const mockState: MockState = {
   previousTrack: vi.fn(),
 };
 
-vi.mock("../../../stores/audioStore", () => ({
+vi.mock("../../../shared/stores/audioStore", () => ({
   useAudioStore: () => mockState,
 }));
