+++ web/apps/mfe-spectrogram/src/components/__tests__/PlaybackControls.test.tsx
@@ -118,7 +118,7 @@ vi.mock("@/shared/stores/uiStore", () => ({
   }),
 }));
 
-vi.mock("../../stores/settingsStore", () => ({
+vi.mock("../../shared/stores/settingsStore", () => ({
   useSettingsStore: () => ({
     theme: "dark",
     amplitudeScale: "db",
