+++ web/apps/mfe-spectrogram/src/components/spectrogram/SpectrogramView.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState, useCallback, useMemo } from 'react'
-import { useAudioStore } from '@/stores/audioStore'
+import { useAudioStore } from '@/shared/stores/audioStore'
 import { useSettingsStore } from '@/shared/stores/settingsStore'
 import { useAudioFile } from '@/hooks/useAudioFile'
 import { useMicrophone } from '@/hooks/useMicrophone'
