+++ web/apps/mfe-spectrogram/src/shared/stores/__tests__/audioPreferences.test.ts
