--- web/apps/mfe-spectrogram/src/stores/playlistSearchStore.ts
@@ -1,11 +1,0 @@
-import { create } from 'zustand'
-
-interface PlaylistSearchState {
-  searchQuery: string
-  setSearchQuery: (query: string) => void
-}
-
-export const usePlaylistSearchStore = create<PlaylistSearchState>((set) => ({
-  searchQuery: '',
-  setSearchQuery: (searchQuery) => set({ searchQuery })
-}))
