--- web/apps/mfe-spectrogram/src/stores/settingsStore.ts
@@ -1,7 +1,0 @@
-/** Thin re-export to keep legacy import paths functional. */
-export {
-  useSettingsStore,
-  STORAGE_KEY,
-  VALID_SEEKBAR_MODES,
-} from "../shared/stores/settingsStore";
-
